fix: resolve merge conflicts and compile errors from chimahon sync
Backup: use BackupExtensionRepos for anime repos, BackupExtensionStore for manga stores. Remove backupExtensionRepo field from Backup class, migrate LegacyBackup conversion. Add mergeExtensionStoreLists for BackupExtensionStore sync. Add missing awaitSuccess, BackupExtensionRepos, BackupExtensionStore imports. Fix backupChapterMapper lambda with explicit 17-param signature including ocrReady. Remove broken duplicate backupAppPreferences stub in BackupCreator. Fix collectAsState types for Flow<Long> (0L) and add Compose import alongside Preference import. Add 14 missing extension repo strings to i18n. Fix SettingsAdvancedScreen DELEGATED_SOURCES .values (List not Map). Fix AppModule UpdateStrategyColumnAdapter -> MangaUpdateStrategyColumnAdapter. Uncomment verboseLogging in NetworkPreferences. Fix MangaExtensions.kt with hasCustomCover import and inline komikku API. Clean up DatabaseAdapter extra brace and ChapterRepositoryImpl dead mapChapter. Add ExtensionReposScreen components and BackupExtensionRepos model from merge.

+++ app/src/main/java/eu/kanade/presentation/more/settings/screen/SettingsBrowseScreen.kt
@@ -26,6 +26,7 @@ import tachiyomi.i18n.kmk.KMR
 import tachiyomi.i18n.sy.SYMR
 import tachiyomi.presentation.core.i18n.pluralStringResource
 import tachiyomi.presentation.core.i18n.stringResource
+import androidx.compose.runtime.collectAsState
 import tachiyomi.presentation.core.util.collectAsState
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
@@ -47,8 +48,8 @@ object SettingsBrowseScreen : SearchableSettings {
         val getExtensionStoreCountAsFlow = remember { Injekt.get<GetExtensionStoreCountAsFlow>() }
         val getAnimeExtensionRepoCount = remember { Injekt.get<GetAnimeExtensionRepoCount>() }
 
-        val reposCount by getExtensionStoreCountAsFlow().collectAsState(0)
-        val animeReposCount by getAnimeExtensionRepoCount.subscribe().collectAsState(0)
+        val reposCount by getExtensionStoreCountAsFlow().collectAsState(0L)
+        val animeReposCount by getAnimeExtensionRepoCount.subscribe().collectAsState(0L)
 
         // SY -->
         val scope = rememberCoroutineScope()
@@ -153,7 +154,7 @@ object SettingsBrowseScreen : SearchableSettings {
                     ),
                     Preference.PreferenceItem.TextPreference(
                         title = "${stringResource(MR.strings.label_anime)} ${stringResource(MR.strings.label_extension_repos)}",
-                        subtitle = pluralStringResource(MR.plurals.num_repos, animeReposCount, animeReposCount),
+                        subtitle = pluralStringResource(MR.plurals.num_repos, animeReposCount.toInt(), animeReposCount),
                         onClick = {
                             navigator.push(AnimeExtensionReposScreen())
                         },
